% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gafunction.R
\name{calculateGAI}
\alias{calculateGAI}
\title{Calculate GA-I}
\usage{
calculateGAI(array1, array2)
}
\arguments{
\item{array1}{array containing the progenitor's genotypes}

\item{array2}{array containing the sample's genotypes}
}
\description{
This function calculates the GA-I score of a sample.
}
