% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gafunction.R
\name{calculateGAIII}
\alias{calculateGAIII}
\title{Calculate GA-III}
\usage{
calculateGAIII(ploidy, array1, array2, array3)
}
\arguments{
\item{ploidy}{population ploidy value, 4 for tetraploid and 6 for hexaploid.}

\item{array1}{array containing the progenitor1's genotypes.}

\item{array2}{array containing the progenitor2's genotypes.}

\item{array3}{array containing the sample's genotypes}
}
\description{
This function calculates the GA-III score of a sample.
}
